refactor(CarRaceController): 게임 주요 로직 따로 분리
